Now average by images processed, not images found
If the user tried to cancel the process halfway through, we would save
the image, but we would average on the number of files found to
process, not the potentially shorter number of images actually
processed, resulting in a darker image.

Also, if a user cancels at just at the right time, we'd throw a
KeyboardInterrupt, now we dont.